move processDeclarationFunctions to stylable-utils
